Create a basic model migration tool from CellML 1.1 to my CellML 1.2 draft.
It currently has the rather major limitation that you have to call the variable of integration the same thing in every component, and specify that name on the command line. This is a difficult issue to fix in general, especially when writing a tool that works on individual XML files, rather than the model as a whole, because the correct variable might not be present in the file we are processing at all.

However, it might be possible to detect some of the easier to handle cases automatically in the future, and only fall back to requiring manual fixups in the harder (and presumably rare) cases.

Note that there is also an issue with CellML models that use initial_value without specifying a rate to mean 'constant'. That case is not handled well by the program.